+++ src/FluentDraft/Messages/UpdateAvailableMessage.cs
@@ -1,0 +1,14 @@
+using Velopack;
+
+namespace FluentDraft.Messages
+{
+    public class UpdateAvailableMessage
+    {
+        public UpdateInfo Update { get; }
+
+        public UpdateAvailableMessage(UpdateInfo update)
+        {
+            Update = update;
+        }
+    }
+}