+++ src/FluentDraft/Messages/UpdateProgressMessage.cs
@@ -1,0 +1,12 @@
+namespace FluentDraft.Messages
+{
+    public class UpdateProgressMessage
+    {
+        public int Progress { get; }
+
+        public UpdateProgressMessage(int progress)
+        {
+            Progress = progress;
+        }
+    }
+}